maintain_kubeusers: Start the script up
This initial commit includes code that genuinely works to create a mock
user with x509 auth on a local system with minikube.  The code currently
works with current versions of the kubernetes python client (10.0.0) and
is tested with the buster package version as well (7.0.0) in case we
continue later with a host-based installation or a buster container with
no pip for some reason.

TODOs are everywhere still, and I'm hoping to add tests, Dockerfile,
k8s service definition for it, and a simple script to add this service's
credentials as well (it shouldn't run as root if it doesn't have to).  It's container
*would* have to run as root to connect to the NFS volume claim to create
homedirs and such.

The working parts need cleanup and review.  It also needs more testing
since the edit permissions given to users must be blocked on volumes by PSP.
If PSP doesn't block that enough, then we need a custom role that removes
volumes and volume claims :)  It's likely we'll end up with a custom role anyway,
but this was sufficient for PoC at least.

Bug: T228499
Change-Id: I2e292cfce7b5f3e2bbdda79d7167dbd1ebbd7814