maintain_kubeusers: add test suite
This adds tests to the repo for use with CI or manually.
They can be run by using pip in your venv to install pytest and pytest-vcr
and running py.test.  That is all handled by running "tox" however, if you
simply install that.  It will then run as it will in CI.

Bug: T228499
Change-Id: Id7fc4c7270b7819bb2aba16aebe144423053a236