maintain_kubeusers: authenticate in k8s and out -- fix tests and docker
This is much closer to actually running in k8s.  I want to set up a live
test, so I may need a local ldap to do it (or just a mockup function).

Overall, this now launches in Kubernetes (if build from the Dockerfile),
will have permissions to do as it needs, and can be run locally with admin
on the commandline.  This also fixes the auth for the tests so that there
is dummy creds that work with the vcr tape mocks of K8s API calls.

Bug: T228499
Change-Id: I90c48e5f83f013c96cc50bd20dd35aa3ea45c678