live-testing: make this work inside a k8s cluster
These changes to the service definition include sufficient permissions.
It still will need to mount the persistent volume claim correctly and bind
to the clusterrole that will be used for Toolforge tools.

This also includes the dynamically produced PSPs needed to make everything
behave in kubernetes for tools as it does now.

Commented in there is a liveness probe that will restart things if they
hang.

Also, add the receive_timeout value that we wish we had in the current
version.

The cassettes are rewritten for the test suite to work inside minikube
and outside it.

Bug: T228499
Change-Id: I8397061648e27822429a74d537c64307541de735