merging: merge configs when one already exists
This should create a "toolforge" context alongside the "default" one that
is created by maintain_kubeusers 1.0. The service.yaml for tools will run
the app with the --gentle-mode option that will set the old cluster as the default.

This approach adds access to EVERY tool.  They just won't actually use it
unless they run "kubectl config use-context toolforge".

In addition, this won't work unless we make the config writeable (or
kubectl cannot change the context).  It isn't right now.
I have verified that the format produced by this is just fine with old
toolforge via live testing on a useless tool of mine.

This also corrects a typo in the config that was introduced at some point.
I validated in minikube that this makes a working config.

Also, adding flock controls to make the write action more polite. NFSv4
expires locks in ~90s so, even in bad situations that should be safe.

Bug: T236202
Change-Id: Id2adcd74b7415648b3bafe17980323bdef3a05b8