migration: add option to switch all remaining users to new cluster
This option is basically just a manipulation of config files.
Any user who is still using the old "default" profile is switched to
"toolforge" (the new cluster).

To that end, we break before checking for new/expiring users, etc. This
only operates on existing users of the cluster, and it should be
mutually exclusive of the option "--gentle-mode" which is the opposite of
this.

Bug: T236202
Bug: T228499
Change-Id: Ice82cb5dda361bd788972b5b33a8ae4af88383f6