testing: add tests for migrating users and make user tests better
This moves the setup parts of the old user test into the test_user fixture
like it should be. This allows writing lots of simple user method tests
such as this one that validates that users are created in gentle mode
and migrated successfully to the toolforge version with the switch_context()
method.

This also might make it easier for folks to see how to write more tests.

Bug: T236202
Bug: T228499
Change-Id: Ie359209007daaeebde23a1479135697482f6a71c