psp: add drop of all capabilities
Tested that webservice still runs with this. I think we can
get away with dropping all capabilties in tool containers.

Tests still run alright, but it would be good to add a refresh to the
vcr cassettes to make the objects match better. That doesn't need
to happen right away.

Bug: T240009
Change-Id: Icffb2ae8337aae09289afb17dfaafe36ddb04915