users: filter out any invalid tool names
Tool names are namespaces and repeated all over Kubernetes. Therefore,
they must meet the definition of an RFC 1123 DNS label, and a period in
a name will cause issues in places. Therefore, maintain_kubeusers will
simply ignore the tools we currently have that are not up to code.

This will silently filter them so that we no longer get this service
misreporting how many tools it created in the log.

Bug: T176027
Bug: T176681
Bug: T240922
Change-Id: I0e7bb9ae56fd79801e5e1a6d13b6daa94e943fd5