kustomize: refactor the yaml for launching the deployment
This separates out files a little for organization, but more importantly
only the changed resource requires new files. This has been tested locally.

Also updated the documentation for deploy in the README and corrected
the testing information a bit.

A benefit of this is that it makes it simple to patch over the base
for deploying a service like this for PAWS to hand out admin credentials.

Bug: T211096
Change-Id: I20878d92c1a3161a0b311c2cbd924b9dbd62e8f3