rbac: move the toolforge roles out of puppet and into here
This code is what connects this role to tools, and the clusterrole is
required for this to run in a cluster. Therefore, it should be here
rather than in puppet.

Placing it in the deployment-base will deploy it in test, beta and
toolforge. Also including the privileged PSP in the test deployment
so we don't actually need puppet at all.

Change-Id: I6f8dd2fd5bd861c4aa2f69ec36d9864bd9355d2b