tests: fix cassette generation and add some testing
This fixes the tests so that there is a proper teardown for API activity.
That will eliminate the need for the fail and retest steps when regenerating
the cassettes.

Adding type hinting to files as I change them and the mypy tests to
check them as well as enforcing the black formatting in tox.

This is notably an effort to remove the testing changes from
Ibe8c6c160ffe410c38a0 so that they will help with development on things
like the attached bug.

Bug: T247455
Change-Id: Ibd6b2a299d9902a58280b1cf8b772b4b584c0d22