kustomize: Make the yaml more declarative
Since I wasn't patching the deployments, I was replacing them,
it seems better to define the deployment entirely for each environment
unless we actually shift to patching -- which would require redefining
the entire container spec.

Overall, this would allow us to see a reassuring "unchanged" message on
the deployment for "kubectl apply". I think that might be good. A patch
*always* says "configured". That is ok if it really is a patch. It just
isn't.

Change-Id: I697045cee4ac1a2834c9e05d655dafc8c610365a