operations pods: update the ops pod yaml and document the whole message
I've been using this for introspecting the cluster for a while, but
I never documented it. It's not a perfect setup, and it would be good
to improve it in time (possibly adding it to the deployments), but it
is better to have it documented and not include breaking yaml!

Change-Id: I0fa27035d5dfd395717b07bfaa58c1bf1a5d0782