crashes: retry creating CSRs after crashes
There have been at least 2 instances of a crash of maintain_kubeusers
where a CSR was left behind. The crashes were motivated by etcd slowness,
but importantly, after a crash, the service should recover. An existing
CSR would send this into a crash/restart loop and end the service.

With this, if the API returns 409, the object is deleted and the method
calls itself. That should allow things to progress since we do not keep
CSRs around in general.

Bug: T271847
Change-Id: I979d5760d485fa0ae78b407ffdb78e1c51f07ece