kubeconfig: use relative paths for certs
With a relative path, makes the home of the user relocatable, so
kubectl (and any other command using the kubeconfig certs) is able to
run no matter the path to the kubeconfig file.

This might only affects new users, old users will retain the old full
paths until the configs are regenerated.

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Icc3cc3244cf72f0f8cd411c68a1af88dde618db6